Bug fix and update
In the previous version band-pass filtered data was unintentionally hilbert transformed twice. This bug is know fixed.
In the case where a parcellation is used that differs substantially in terms of number of parcels compared to the one used in the original manuscript, the OV parameter (used for the grouping of SNCs into SNs) should be set heuristically. This is now implemented in the new version of the code.